tango2/requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.1
django-mysql==3.8.1
dotmap==1.3.14
fpdf==1.7.2
future==0.18.2
html5lib==1.1
mysqlclient==2.0.1
numpy==1.18.5
pandas==1.0.4
pathlib==1.0.1
Pillow==7.1.2
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
reportlab==3.5.53
six==1.15.0
sqlparse==0.3.1
webencodings==0.5.1
xhtml2pdf==0.2.5
